ai_trust_engine/requirements.txt
# SeaSense AI Trust Engine Requirements
# Smart India Hackathon 2025

# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning and AI
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentence-transformers>=2.2.0
git+https://github.com/openai/CLIP.git
scikit-learn>=1.3.0
numpy>=1.21.0,<1.25.0
scipy>=1.9.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=9.0.0
ultralytics>=8.0.0

# NLP Libraries
spacy>=3.4.0
nltk>=3.7.0
langdetect>=1.0.0

# Database and Caching
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
sqlite3  # Built-in with Python

# Data Processing
pandas==2.1.4
aiofiles==23.2.1

# HTTP and API clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Clustering and Similarity
faiss-cpu==1.7.4
annoy==1.17.3

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
typer==0.9.0
click==8.1.7

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and Observability
prometheus-client==0.19.0
structlog==23.2.0

# Social Media APIs (Optional)
tweepy==4.14.0
facebook-sdk==3.1.0

# Geospatial
geopy==2.4.1
geopandas==0.14.1
shapely==2.0.2

# Security
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# Environment and Configuration
python-decouple==3.8
pyyaml==6.0.1

# Background Tasks
celery==5.3.4
kombu==5.3.4

# WebSockets (for real-time features)
websockets==12.0

# Data Validation and Serialization
marshmallow==3.20.1
jsonschema==4.20.0

# Timing and Scheduling
apscheduler==3.10.4
croniter==2.0.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Performance
cachetools==5.3.2
async-lru==2.0.4
